+++ skgstat/tests/test_variogram.py
@@ -588,7 +588,7 @@ class TestVariogramQaulityMeasures(unittest.TestCase):
         ):
             V.set_model(model)
             self.assertAlmostEqual(V.r, r, places=2)
-    
+
     def test_NS(self):
         V = Variogram(self.c, self.v, n_lags=15, normalize=False)
 
@@ -597,6 +597,34 @@ class TestVariogramQaulityMeasures(unittest.TestCase):
             [0.0206, 0.0206, 0.0206]
         ):
             self.assertAlmostEqual(V.NS, NS, places=4)
+
+    def test_update_kwargs(self):
+        V = Variogram(self.c, self.v, percentile=.3)
+
+        self.assertAlmostEqual(
+            V._kwargs.get('percentile'), 0.3, places=1
+        )
+
+        # change the parameter
+        V.update_kwargs(percentile=0.7)
+
+        self.assertAlmostEqual(
+            V._kwargs.get('percentile'), 0.7, places=1
+        )
+
+    def test_kwargs_setter_in_experimental(self):
+        V = Variogram(self.c, self.v, estimator='percentile')
+
+        # store with p of 50 == median
+        exp = V.experimental
+
+        V.update_kwargs(percentile=25)
+
+        exp2 = V.experimental
+
+        # 25% should be very different from median
+        with self.assertRaises(AssertionError):
+            assert_array_almost_equal(exp, exp2, decimal=2)
 
 
 class TestVariogramMethods(unittest.TestCase):